+++ UIKitSwagger/View/SwaggerView/SwaggerView.swift
@@ -24,6 +24,7 @@ public class SwaggerView: UIView {
         backgroundColor = nil
     }
 
+
     override public class func layerClass() -> AnyClass {
         return CAShapeLayer.self
     }
@@ -31,6 +32,7 @@ public class SwaggerView: UIView {
     private var shapeLayer: CAShapeLayer {
         return layer as! CAShapeLayer
     }
+
 
     override public var backgroundColor: UIColor? {
         get {
@@ -44,5 +46,8 @@ public class SwaggerView: UIView {
             shapeLayer.fillColor = newValue?.CGColor
         }
     }
+
+
+    @IBInspectable public var cornerRadius: CGFloat = 0
 
 }